Added compatibility class for 2.x version
