+++ src/HTML5.php
@@ -1,0 +1,7 @@
+<?php
+namespace Masterminds;
+
+class HTML5 extends AbstractHtml5
+{
+
+}
